--- src/main/java/BS/QACinema/dataLayer/BooksSeats/BookASeat.java
@@ -1,11 +1,0 @@
-/*//
- * Created by Martyn Green
-*/
-package BS.QACinema.dataLayer.BooksSeats;
-
-public class BookASeat {
-	
-	private void checkNextUnbookedSeat(Screen Screen, String time){
-		//getFromDB.NextUnbookedSeat();
-	}
-}
