--- src/main/java/BS/QACinema/dataLayer/booking/tickets/CustomerType.java
@@ -1,5 +1,0 @@
-package BS.QACinema.dataLayer.booking.tickets;
-
-public enum CustomerType {
-	CHILD, OAP, ADULT, STUDENT, ALL
-}
