--- src/main/java/BS/QACinema/dataLayer/booking/tickets/DaysOfWeek.java
@@ -1,5 +1,0 @@
-package BS.QACinema.dataLayer.booking.tickets;
-
-public enum DaysOfWeek {
-	MONDAY, TUESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY, ALL, WEEKDAYS, WEEKNDS
-}
