--- src/main/java/BS/QACinema/dataLayer/booking/tickets/TimesOfDay.java
@@ -1,5 +1,0 @@
-package BS.QACinema.dataLayer.booking.tickets;
-
-public enum TimesOfDay {
-	MORNING, AFTERNOON, ALL_DAY
-}
