--- src/main/java/BS/QACinema/dataLayer/users/User.java
@@ -1,5 +1,0 @@
-package BS.QACinema.dataLayer.users;
-
-public class User {
-
-}
