+++ src/main/java/qa_cinema/admin/Placeholder.java
@@ -1,0 +1,5 @@
+package qa_cinema.admin;
+
+public class Placeholder {
+
+}
