+++ src/main/java/qa_cinema/data/booking/BookASeat.java
@@ -1,0 +1,13 @@
+/*//
+ * Created by Martyn Green
+*/
+package qa_cinema.data.booking;
+
+import qa_cinema.data.cinema.Screen;
+
+public class BookASeat {
+	
+	private void checkNextUnbookedSeat(Screen Screen, String time){
+		//getFromDB.NextUnbookedSeat();
+	}
+}
