+++ src/main/java/qa_cinema/data/booking/tickets/CustomerType.java
@@ -1,0 +1,5 @@
+package qa_cinema.data.booking.tickets;
+
+public enum CustomerType {
+	CHILD, OAP, ADULT, STUDENT, ALL
+}
