+++ src/main/java/qa_cinema/data/booking/tickets/DaysOfWeek.java
@@ -1,0 +1,5 @@
+package qa_cinema.data.booking.tickets;
+
+public enum DaysOfWeek {
+	MONDAY, TUESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY, ALL, WEEKDAYS, WEEKNDS
+}
