+++ src/main/java/qa_cinema/data/booking/tickets/TimesOfDay.java
@@ -1,0 +1,5 @@
+package qa_cinema.data.booking.tickets;
+
+public enum TimesOfDay {
+	MORNING, AFTERNOON, ALL_DAY
+}
