+++ src/main/java/qa_cinema/data/film/Genre.java
@@ -1,0 +1,37 @@
+package qa_cinema.data.film;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import javax.validation.constraints.NotNull;
+
+@Entity @Table
+public class Genre {
+
+	@Id
+	@NotNull
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	private int id;
+	
+	@NotNull
+	private String genre;
+	
+	
+	public Genre(String genre){
+		this.genre = genre;
+	}
+	
+	public String getGenre() {
+		return genre;
+	}
+
+	public void setGenre(String genre) {
+		this.genre = genre;
+	}
+	
+	public int getId(){
+		return id;
+	}
+}
