+++ src/main/java/qa_cinema/data/users/User.java
@@ -1,0 +1,5 @@
+package qa_cinema.data.users;
+
+public class User {
+
+}
