--- src/test/java/BS/QACinema/dataLayer/RoleTest.java
@@ -1,32 +1,0 @@
-package BS.QACinema.dataLayer;
-
-import BS.QACinema.dataLayer.film.*;
-import static org.junit.Assert.*;
-
-import org.junit.Before;
-import org.junit.Test;
-
-public class RoleTest {
-	
-	Role rol;
-	
-	@Before
-	public void setUp() {
-		rol = new Role("MrsTest");
-	}
-	
-	@Test
-	public void testGetRole(){
-		
-		assertEquals("MrsTest", rol.getRole());
-	}
-	
-	@Test
-	public void testSetRole(){
-		
-		String newRole = "MrTest";
-		rol.setRole(newRole);
-		assertEquals("MrTest", rol.getRole());
-	}
-
-}
